refactor(nameReducer): refactor the name reducer into a slice
