+++ src/hooks/names.ts
@@ -1,18 +1,18 @@
 import { useMemo, useCallback } from 'react'
 import { useSelector } from 'react-redux'
 
-import { getNames, getAvatarNameById } from '../selectors/names'
+import { selectNames, selectAvatarNameById } from '../reducers/names'
 import { getGroups } from '../selectors/groups'
 
 export function useNames (...ids: string[]) {
-  const names = useSelector(getNames)
+  const names = useSelector(selectNames)
 
   // eslint-disable-next-line
   return useMemo(() => ids.map(id => names[id]), [names, ...ids])
 }
 
 export function useName (id: string) {
-  const selector = useCallback(state => getAvatarNameById(state, id), [id])
+  const selector = useCallback(state => selectAvatarNameById(state, id), [id])
 
   return useSelector(selector)
 }
